tests/requirements.txt
nose
django-nose
coverage